+++ src/Index/UnitTest/Mocks/EmptyTermTable.h
@@ -19,11 +19,11 @@ namespace BitFunnel
         // These methods are not applicable for the tests and throw if called.
         virtual void Write(std::ostream& stream) const override;
         virtual void AddRowId(RowId id) override;
-        virtual unsigned GetRowIdCount() const override;
-        virtual RowId GetRowId(unsigned rowOffset) const override;
-        virtual RowId GetRowIdAdhoc(uint64_t hash, unsigned rowOffset, unsigned variant)  const override;
-        virtual RowId GetRowIdForFact(unsigned rowOffset) const override;
-        virtual void AddTerm(uint64_t hash, unsigned rowIdOffset, unsigned rowIdLength) override;
+        virtual size_t GetRowIdCount() const override;
+        virtual RowId GetRowId(size_t rowOffset) const override;
+        virtual RowId GetRowIdAdhoc(uint64_t hash, size_t rowOffset, size_t variant)  const override;
+        virtual RowId GetRowIdForFact(size_t rowOffset) const override;
+        virtual void AddTerm(uint64_t hash, size_t rowIdOffset, size_t rowIdLength) override;
         /*
         virtual void AddTermAdhoc(Stream::Classification classification,
                                   unsigned gramSize,
@@ -33,8 +33,8 @@ namespace BitFunnel
                                   unsigned rowIdLength) override;
         */
         virtual void SetRowTableSize(size_t rank,
-                                     unsigned rowCount,
-                                     unsigned sharedRowCount) override;
+                                     size_t rowCount,
+                                     size_t sharedRowCount) override;
         virtual size_t GetMutableFactRowCount(size_t rank) const override;
         virtual PackedTermInfo GetTermInfo(const Term& term, TermKind& termKind) const override;
